evas-gl-drm: Cleanup compiler warnings when building gl_drm engine
Summary: This cleans up nasty compiler warnings when building the
gl_drm engine. If we do not define EGL_EGLEXT_PROTOTYPES then we end
up with implicit declaration warnings

@fix

+++ src/modules/evas/engines/gl_drm/evas_outbuf.c
@@ -246,7 +246,7 @@ _evas_outbuf_egl_setup(Outbuf *ob)
         DBG("Config Format: %d", format);
         DBG("OB Format: %d", ob->info->info.format);
 
-        if (format == ob->info->info.format)
+        if (format == (int)ob->info->info.format)
           {
              ob->egl.config = cfgs[i];
              break;
